# add new float-parser based on eisel-lemire algorithm 