--- a/lib/std/fmt.zig
+++ b/lib/std/fmt.zig
@@ -1837,8 +1837,8 @@ test "parseUnsigned" {
 }
 
 pub const parseFloat = @import("fmt/parse_float.zig").parseFloat;
+pub const parseHexFloat = @compileError("deprecated; use `parseFloat`");
 pub const ParseFloatError = @import("fmt/parse_float.zig").ParseFloatError;
-pub const parseHexFloat = @import("fmt/parse_hex_float.zig").parseHexFloat;
 
 test {
     _ = parseFloat;