@tiehuis 2022-05-02
I expect adding support for `f80` in future should be fairly straight-forward (slow-method) since there should be no major internal changes now required except for handling the explicit bit.