@andrewrk 2022-05-02
I had a look at the failures in your branch - most of the problems are solved by 9af4cada73e43fb98e906a0ed79197963d1a05ac which I just pushed to master. There was one remaining problem which is `test "quad hex float literal parsing accurate"` from test/behavior/math.zig. This one could probably use your expertise.