@tiehuis 2022-05-02
So this specific test-case looks incorrect. `0x1.edcb34a235253948765432134674fp-1` should be represented as `0x1.edcb34a235253948765432134675p-1`, rounding up and not `0x1.edcb34a235253948765432134674p-1` which is what is incorrectly parsed by the current hex float parser. This is obvious given the last truncated digit that is used in the rounding process (`f`).

I've included a small script as well to see the difference between the f128 values which bound the true value, in decimal form:

```python
from decimal import *
import math

getcontext().prec = 64

af = "0x1.edcb34a235253948765432134674fp-1"
accurate = Decimal(0.1) * (Decimal(1) +
    Decimal(0xedcb34a235253948765432134674f) /
    Decimal(0x100000000000000000000000000000))


lf = "0x1.edcb34a235253948765432134674p-1"
lo = Decimal(0.1) * (Decimal(1) +
    Decimal(0xedcb34a235253948765432134674) /
    Decimal(0x10000000000000000000000000000))

hf = "0x1.edcb34a235253948765432134675p-1"
hi = Decimal(0.1) * (Decimal(1) +
    Decimal(0xedcb34a235253948765432134675) /
    Decimal(0x10000000000000000000000000000))

print(hf, abs(accurate-hi))
print(lf, abs(accurate-lo))
```

```
('0x1.edcb34a235253948765432134675p-1', Decimal('1.2037062152420224749791039664E-36'))
('0x1.edcb34a235253948765432134674p-1', Decimal('1.80555932286303371246865594964E-35'))
```

The issue looks to have been introduced here: https://github.com/ziglang/zig/commit/7f024d6786b690e7980d6e47d7ae70fa8528a0df

Specifically the check `if (guardBit and (exactly_half or more_than_half)) {`. It looks like this should have been `if ((guardBit and exactly_half) or more_than_half) {` based on the comments about how the rounding process was meant to function. Modifying the existing parser with these changes makes them agree. The current rounding behavior appears to be round-up, only for odd values.

The "fix" for this PR (that I will test later) should be modifying the test-case to use the correctly rounded bit representation:

```
try expect(@bitCast(u128, f) == 0x3ffeedcb34a235253948765432134675);
```