@andrewrk 2022-05-03
Looks like the failure is this one:

```
/workspace/lib/std/fmt/parse_float.zig:43:9: 0x5887a7 in fmt.parse_float.test "std-aarch64-linux-none-Debug-bare-multi fmt.parseFloat" (test)
        try expectEqual(@bitCast(Z, try parseFloat(T, "nAn")), @bitCast(Z, std.math.nan(T)));
        ^
```

You should be able to reproduce it like this:

```
zig test lib/std/std.zig -target aarch64-linux -femit-bin=test
qemu-aarch64 ./test
```